requirements.txt
# Requirements for MonoSplat project with Nersemble dataset support
# Generated from conda environment: /data/baosongze/env/splatter_image
# Date: 2026-01-19
#
# Note: Some packages are installed from local paths and may need special handling:
# - diff_gaussian_rasterization @ file:///data/baosongze/workspace/3D/splatter-image/gaussian-splatting/submodules/diff-gaussian-rasterization
# - simple_knn @ file:///data/baosongze/workspace/3D/splatter-image/gaussian-splatting/submodules/simple-knn
#
# To install: pip install -r requirements.txt
# For local packages, you may need to install them separately.

aiofiles==23.2.1
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
altair==5.5.0
antlr4-python3-runtime==4.9.3
anyio==4.12.1
arrow==1.4.0
async-timeout==5.0.1
attrs==25.4.0
beartype==0.22.9
beautifulsoup4==4.14.3
black==25.12.0
blessed==1.25.0
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
colorspacious==1.1.2
contourpy==1.3.2
croniter==1.3.15
cycler==0.12.1
dacite==1.9.2
dateutils==0.6.12
decorator==4.4.2
deepdiff==7.0.1
diff_gaussian_rasterization @ file:///data/baosongze/workspace/3D/splatter-image/gaussian-splatting/submodules/diff-gaussian-rasterization
e3nn==0.5.9
editor==1.6.6
einops==0.8.1
ema-pytorch==0.7.9
exceptiongroup==1.3.1
fastapi==0.88.0
ffmpy==1.0.0
filelock==3.20.3
fonttools==4.61.1
frozenlist==1.8.0
fsspec==2023.12.2
gitdb==4.0.12
GitPython==3.1.46
gradio==3.50.2
gradio_client==0.6.1
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.3.1
hydra-core==1.3.2
idna==3.11
ImageIO==2.37.2
imageio-ffmpeg==0.6.0
importlib_resources==6.5.2
inquirer==3.4.1
itsdangerous==2.2.0
jaxtyping==0.3.5
Jinja2==3.1.6
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
lazy_loader==0.4
lightning==2.6.0
lightning-cloud==0.6.0
lightning-utilities==0.15.2
lpips==0.1.4
markdown-it-py==4.0.0
MarkupSafe==2.0.1
matplotlib==3.10.8
mdurl==0.1.2
moviepy==1.0.3
mpmath==1.3.0
multidict==6.7.0
mypy_extensions==1.1.0
narwhals==2.15.0
networkx==3.4.2
numpy==2.2.6
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu12==12.8.90
omegaconf==2.3.0
opencv-python==4.12.0.88
opt-einsum-fx==0.1.4
opt_einsum==3.4.0
ordered-set==4.1.0
orjson==3.11.5
packaging==24.2
pandas==2.3.3
pathspec==1.0.3
pillow==10.4.0
platformdirs==4.5.1
plyfile==1.1.3
proglog==0.1.12
propcache==0.4.1
protobuf==6.33.3
psutil==6.1.1
pydantic==1.10.26
pydub==0.25.1
Pygments==2.19.2
pyparsing==3.3.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.21
pytokens==0.3.0
pytorch-lightning==2.6.0
pytz==2025.2
PyYAML==6.0.3
readchar==4.2.1
referencing==0.37.0
requests==2.32.5
rich==14.2.0
rpds-py==0.30.0
ruff==0.14.13
runs==1.2.2
safetensors==0.7.0
scikit-image==0.25.2
scipy==1.15.3
semantic-version==2.10.0
sentry-sdk==2.49.0
shellingham==1.5.4
simple_knn @ file:///data/baosongze/workspace/3D/splatter-image/gaussian-splatting/submodules/simple-knn
six==1.17.0
sk-video==1.1.10
smmap==5.0.2
soupsieve==2.8.1
starlette==0.22.0
starsessions==1.3.0
svg.py==1.10.0
sympy==1.14.0
tabulate==0.9.0
tifffile==2025.5.10
timm==1.0.24
tomli==2.4.0
torch==2.9.1
torchmetrics==1.8.2
torchvision==0.24.1
tqdm==4.67.1
traitlets==5.14.3
triton==3.5.1
typer-slim==0.21.1
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
uvicorn==0.40.0
wadler_lindig==0.1.7
wandb==0.23.1
wcwidth==0.2.14
websocket-client==1.9.0
websockets==11.0.3
xmod==1.8.1
yarl==1.22.0